refactor(device_api): extract DevApiBackend vtable, split default/nvshmem backends
