+++ test/unittest/shmem_adaptor/Makefile
@@ -1,20 +1,21 @@
-# Build: make NVSHMEM_HOME=/path/to/nvshmem CUDA_HOME=/usr/local/cuda
+# Build: make SHMEM_HOME=/path/to/nvshmem CUDA_HOME=/usr/local/cuda
 # Run:   nvshmrun -np 2 ./test_nvshmem_adaptor
 
-NVSHMEM_HOME ?= /usr/local/nvshmem
+SHMEM_HOME ?= /usr/local/nvshmem
 CUDA_HOME ?= /usr/local/cuda
 
 NVCC := $(CUDA_HOME)/bin/nvcc
 NVCC_FLAGS := -std=c++17 -O2 -g \
-	-DFLAGCX_COMM_TRAITS_NVSHMEM -DUSE_NVIDIA_ADAPTOR \
+	-DFLAGCX_COMM_TRAITS_SHMEM -DUSE_NVIDIA_ADAPTOR \
 	-I../../../flagcx/include \
 	-I../../../flagcx/adaptor/include \
+	-I../../../flagcx/adaptor/device_api \
 	-I../../../flagcx/adaptor/shmem \
-	-I$(NVSHMEM_HOME)/include \
+	-I$(SHMEM_HOME)/include \
 	-I$(CUDA_HOME)/include \
 	-rdc=true
 
-LDFLAGS := -L$(NVSHMEM_HOME)/lib -lnvshmem_device -lnvshmem_host \
+LDFLAGS := -L$(SHMEM_HOME)/lib -lnvshmem_device -lnvshmem_host \
 	-L$(CUDA_HOME)/lib64 -lcudart -lcuda
 
 TARGET := test_nvshmem_adaptor
